Support any authentication strategy using ACLPermissionOverride extension point.
This uses the new ACLPermissionOverride ExtensionPoint in Jenkins to let us delegate wrap the chosen strategy with some exceptions.

We support the boolean options via the GithubSecurityRealm and then apply them ahead of the chosen AuthorizationStrategy.

This way users can use the Matrix strategies and improvements in those controls are transparently available to